+++ android/src/main/java/io/parity/signer/ui/navigationselectors/ScreenSelectorsNew.kt
@@ -2,7 +2,7 @@ package io.parity.signer.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.State
-import io.parity.signer.bottomsheets.exportprivatekey.PrivateKeyExportBottomSheet
+import io.parity.signer.screens.keydetails.exportprivatekey.PrivateKeyExportBottomSheet
 import io.parity.signer.models.*
 import io.parity.signer.screens.keydetails.KeyDetailsMenuAction
 import io.parity.signer.screens.keysets.NewSeedMenu
